Cambios diversos

+++ application/controllers/Users.php
@@ -3,7 +3,7 @@
 include_once('Security.php');
 
 class Users extends Security {
-
+    
     public function view_users() {
         $id = $this->session->userdata("id");
         $nivel = $this->modelUser->getNivel($id);
@@ -23,9 +23,10 @@ class Users extends Security {
     public function insert_user(){
         $usuario = $this->input->get_post('usuarioIns');
         $contrasena = $this->input->get_post('contrasenaIns');
+        $hash = $this->modelUser->hash_pass($contrasena);
         $nivel = $this->input->get_post('nivelIns');
         
-        $r = $this->modelUser->insert($usuario, $contrasena, $nivel);
+        $r = $this->modelUser->insert($usuario, $hash, $nivel);
         
         if ($r == 0) {
             $data["msg"] = "1";
@@ -42,15 +43,21 @@ class Users extends Security {
     }
     
     public function delete_user($id) {
-        $r = $this->modelUser->delete($id);
-        
-        if ($r != 0) {
-            $data["msg"] = "0";
-            $data['ListaUsuarios'] = $this->modelUser->get_all();
-            $data["viewName"] = "admin_users";
-            $this->load->view('template', $data);
+        if ($id != $this->session->userdata("id")) {
+            $r = $this->modelUser->delete($id);
+            if ($r != 0) {
+                $data["msg"] = "0";
+                $data['ListaUsuarios'] = $this->modelUser->get_all();
+                $data["viewName"] = "admin_users";
+                $this->load->view('template', $data);
+            } else {
+                $data["msg"] = "1";
+                $data['ListaUsuarios'] = $this->modelUser->get_all();
+                $data["viewName"] = "admin_users";
+                $this->load->view('template', $data);
+            }
         } else {
-            $data["msg"] = "1";
+            $data["msg"] = "2";
             $data['ListaUsuarios'] = $this->modelUser->get_all();
             $data["viewName"] = "admin_users";
             $this->load->view('template', $data);
@@ -61,9 +68,10 @@ class Users extends Security {
         $id = $this->input->get_post('idMod');
         $usuario = $this->input->get_post('usuarioMod');
         $contrasena = $this->input->get_post('contrasenaMod');
+        $hash = $this->modelUser->hash_pass($contrasena);
         $nivel = $this->input->get_post('nivelMod');
         
-        $r = $this->modelUser->update($usuario, $contrasena, $nivel, $id);
+        $r = $this->modelUser->update($usuario, $hash, $nivel, $id);
         
         if ($r == 0) {
             $data["msg"] = "1";
